# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: CRM ИТ Школы РТК — REST API
  version: 0.2.0
  description: |
    API веб-клиента CRM и точка интеграции с корпоративными системами.

    Аутентификация: вход через Keycloak (GET /api/auth/login) создаёт серверную сессию BFF, браузер получает HttpOnly cookie SESSION; токены OIDC клиенту не передаются. Выход (POST /api/auth/logout) завершает сессию CRM и возвращает адрес завершения сессии Keycloak.

    Изменяющие запросы передают заголовок X-CSRF-TOKEN со значением token из GET /api/csrf. Команды передают Idempotency-Key; повтор с тем же ключом и телом возвращает прежний результат, с другим телом — 409 IDEMPOTENCY_CONFLICT. Устаревшая version даёт 409 VERSION_CONFLICT с currentVersion.

    Ошибки возвращаются в формате ApiError {code, message, requestId}. requestId совпадает с заголовком X-Request-Id и записью в журнале сервера; stack trace и секреты в ответ не попадают.
servers:
  - url: /
...
paths:
...
  /api/reports:
    post:
      tags: [Reports]
      summary: Заказать отчёт или диаграмму
      operationId: createReportJob
      x-implementation-stage: 5
      description: >
        Persists a PENDING report job and hands it to one of the bounded report slots. The same Idempotency-Key with
        the same payload returns the same jobId; a different payload under the key returns 409 IDEMPOTENCY_CONFLICT.
        503 REPORT_CAPACITY_EXCEEDED with Retry-After means all slots and the bounded queue are busy and no job was kept.
        With groupBy the job renders a chart of POST /api/statistics for the same request as PNG or PDF (chart, basis
        table, unit, period, filters and generation time) instead of the row report.
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '202':
          description: Report job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ReportCapacityExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/reports/preview:
    post:
      tags: [Reports]
      summary: Предпросмотр строк отчёта
      operationId: previewReport
      x-implementation-stage: 5
      description: Synchronous page of the same rows that the report file contains, filtered by the current server-side scope
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/Page'
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportPreviewRequest'
      responses:
        '200':
          description: Report rows page with total row count and explanatory notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/statistics:
    post:
      tags: [Reports]
      summary: Статистика для диаграмм
      operationId: reportStatistics
      x-implementation-stage: 5
      description: >
        Counts over exactly the same row set as the report with the same filters, aggregated by the database; rows are
        not loaded into the API process and app.reports.max-rows does not apply. unknownCount holds rows without a
        value for the grouping; groups present in items with count 0 are real zeros (MONTH is zero-filled over the period).
        MONTH with kind PORTFOLIO and periodBasis ACTIVITY is rejected with 400 VALIDATION_ERROR (field groupBy).
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatisticsRequest'
      responses:
        '200':
          description: Grouped counts, total and unknown count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report-jobs:
    get:
      tags: [Reports]
      summary: Последние заказанные отчёты
      operationId: listRecentReportJobs
      x-implementation-stage: 5
      responses:
        '200':
          description: Up to 20 latest report jobs of the current profile that are still valid for its current access revision
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report-jobs/{id}:
    get:
      tags: [Reports]
      summary: Состояние задания отчёта
      operationId: getReportJob
      x-implementation-stage: 5
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Report job state for its owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/ReportGone'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report-jobs/{id}/result:
    get:
      tags: [Reports]
      summary: Скачать файл отчёта
      operationId: downloadReportJobResult
      x-implementation-stage: 5
      description: File is served only to the owner whose role, team and access revision are unchanged since the order
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Report file; Content-Disposition carries a UTF-8 filename*
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/vnd.ms-excel:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: REPORT_NOT_READY, the job has not succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '410':
          $ref: '#/components/responses/ReportGone'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report-filters/managers:
    get:
      tags: [Reports]
      summary: Ответственные для фильтра отчёта
      operationId: listReportManagers
      x-implementation-stage: 5
      responses:
        '200':
          description: USER receives only itself; LEADER receives USER profiles of its team and current owners of visible organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportManagerOption'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report-filters/stages:
    get:
      tags: [Reports]
      summary: Этапы для фильтра отчёта
      operationId: listReportStages
      x-implementation-stage: 5
      responses:
        '200':
          description: Distinct stage names of visible interactions in workflow order
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/admin/sources:
    get:
      tags: [Administration]
      summary: Состояние источников данных
      operationId: listDataSources
      x-implementation-stage: 4
      description: >
        ADMIN only. WEBSITE is the adapter of the proposed site contract (docs/integrations/site-api.yaml), not a
        confirmed API of the customer; MOODLE reads aggregated course and group data through a restricted Moodle web
        service (docs/integrations.md). Base URL, token and the allowed course list come only from the deployment
        configuration and are never returned.
      responses:
        '200':
          description: Sources with configuration flag, last run with counters, last success and number of records to review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataSource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/admin/sources/{source}/sync:
    post:
      tags: [Administration]
      summary: Запустить синхронизацию источника
      operationId: startSourceSync
      x-implementation-stage: 4
      description: >
        ADMIN only. Persists a PENDING run and hands it to the bounded sync executor (not the HTTP thread); returns 202
        with runId. Only one PENDING or RUNNING run per source exists: a second start gets 409 SYNC_ALREADY_RUNNING.
        Repeating a sync is safe: a site record with the same source, type, externalId and updatedAt is skipped, and a
        Moodle snapshot with unchanged counts only gets a new observation time. A source error fails the run and leaves
        stored records, published Moodle snapshots, demand numbers and the updated_since watermark intact.
        The same runs are started by the server schedule SOURCES_SYNC_CRON on behalf of the active administrator who
        started the latest run of the source; a source nobody started yet is not scheduled.
      parameters:
        - name: source
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/SourceCode'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '202':
          description: Sync run accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRunCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: >
            SOURCE_NOT_CONFIGURED when SITE_BASE_URL is empty, or MOODLE_BASE_URL, MOODLE_TOKEN or MOODLE_COURSE_IDS is
            empty; SYNC_ALREADY_RUNNING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: SYNC_CAPACITY_EXCEEDED, the sync executor and its queue are busy; nothing was stored
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/admin/source-records:
    get:
      tags: [Administration]
      summary: Записи источников для разбора
      operationId: listSourceProblemRecords
      x-implementation-stage: 4
      description: ADMIN only. Up to 200 latest records with status NEEDS_MAPPING or FAILED
      responses:
        '200':
          description: Records that need an organization or program mapping, or failed to apply
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/admin/source-records/{id}/apply:
    post:
      tags: [Administration]
      summary: Сопоставить и применить запись источника
      operationId: applySourceRecord
      x-implementation-stage: 4
      description: >
        ADMIN only. For a WEBSITE record saves the optional organization mapping (for the record's external
        organization ID, or its name when the ID is absent) and program mapping (for the program name). For a MOODLE
        NEEDS_MAPPING course or group record organizationId, programId, runStartsOn and runEndsOn are required: the
        mapping confirms a training run [runStartsOn, runEndsOn); a FAILED MOODLE record without them is applied again
        with its saved mapping. A run that ended before its first observation is SKIPPED without a snapshot, a run that
        has ended keeps its last observation. A group cannot be mapped when its course is mapped as a whole and vice
        versa (400). Then applies the stored record again and re-applies up to
        500 other NEEDS_MAPPING records of the same source. Applying an APPLIED or SKIPPED record changes nothing, so
        a repeated request is safe.
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceRecordApply'
      responses:
        '200':
          description: Record after the new attempt and the number of other records applied thanks to the mapping
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceRecordApplyResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
      description: Серверная сессия BFF, созданная входом через GET /api/auth/login. Swagger UI на том же origin использует сессию браузера.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Значение token из GET /api/csrf для текущей сессии. Без заголовка или с чужим значением запрос отклоняется с 403 FORBIDDEN.
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
  responses:
    BadRequest:
      description: Validation or request format error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: BFF session is absent or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Authenticated identity is not allowed to perform the action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    CrmProfileRequired:
      description: The OIDC identity has no active CRM profile (CRM_PROFILE_REQUIRED) or its profile created at first login waits for administrator activation (CRM_PROFILE_PENDING)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Object does not exist or is outside the current server-side scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Version or idempotency conflict; VERSION_CONFLICT includes currentVersion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalError:
      description: Unexpected server error; the stack trace is written only to the server log under the same requestId
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ReportCapacityExceeded:
      description: REPORT_CAPACITY_EXCEEDED, all report slots and the bounded queue are busy; nothing was stored
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ReportGone:
      description: REPORT_ACCESS_CHANGED after a role, team or access revision change, or REPORT_RESULT_UNAVAILABLE when the file is gone
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHENTICATED
            - FORBIDDEN
            - CRM_PROFILE_REQUIRED
            - CRM_PROFILE_PENDING
            - LAST_ACTIVE_ADMIN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VERSION_CONFLICT
            - IDEMPOTENCY_CONFLICT
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - RATE_LIMITED
            - INTERNAL_ERROR
            - DEPENDENCY_UNAVAILABLE
            - REPORT_NOT_READY
            - REPORT_ACCESS_CHANGED
            - REPORT_RESULT_UNAVAILABLE
            - REPORT_CAPACITY_EXCEEDED
            - SOURCE_NOT_CONFIGURED
            - SYNC_ALREADY_RUNNING
            - SYNC_CAPACITY_EXCEEDED
            - LMS_NOT_MAPPED
            - LMS_SYNC_FAILED
        message:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        currentVersion:
          type: integer
          minimum: 0
...
    ReportPreviewRequest:
      type: object
      required: [kind]
      description: from and to are inclusive dates converted to [from 00:00, to+1 00:00) in Europe/Moscow
      properties:
        kind:
          $ref: '#/components/schemas/ReportKind'
        from:
          type: [string, 'null']
          format: date
        to:
          type: [string, 'null']
          format: date
        periodBasis:
          oneOf:
            - $ref: '#/components/schemas/PeriodBasis'
            - type: 'null'
        filters:
          $ref: '#/components/schemas/ReportFilters'
        columns:
          type: array
          description: Omitted or empty means all columns of the kind in the default order
          items:
            $ref: '#/components/schemas/ReportColumn'
        sortBy:
          description: >
            DEMAND only, APPLICATIONS (default), PARTICIPANTS or PARALLEL_RUNS; rows are sorted by that measure
            descending with «нет данных» last, then by program name. Other kinds reject it with 400 VALIDATION_ERROR.
          oneOf:
            - $ref: '#/components/schemas/ReportColumn'
            - type: 'null'
    ReportRequest:
      allOf:
        - $ref: '#/components/schemas/ReportPreviewRequest'
        - type: object
          required: [format]
          properties:
            format:
              $ref: '#/components/schemas/ReportFormat'
            groupBy:
              description: >
                Set to order a chart of the statistics grouped this way instead of the row report; columns are ignored
                and format must be PNG or PDF
              oneOf:
                - $ref: '#/components/schemas/StatisticsGroupBy'
                - type: 'null'
    ReportJobCreated:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: string
          format: uuid
    ReportJob:
      type: object
      required: [id, kind, format, status, progress, resultReady, createdAt]
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/ReportKind'
        format:
          $ref: '#/components/schemas/ReportFormat'
        groupBy:
          description: Grouping of a chart job; null for a row report
          oneOf:
            - $ref: '#/components/schemas/StatisticsGroupBy'
            - type: 'null'
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCEEDED, FAILED]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        resultReady:
          type: boolean
        rowCount:
          type: [integer, 'null']
          minimum: 0
        fileName:
          type: [string, 'null']
        error:
          type: object
          required: [code, message]
          description: Present for FAILED jobs, e.g. REPORT_ROW_LIMIT, REPORT_XLS_ROW_LIMIT, REPORT_PDF_ROW_LIMIT, REPORT_CHART_LIMIT, REPORT_ACCESS_CHANGED, REPORT_INTERRUPTED, REPORT_FAILED
          properties:
            code:
              type: string
            message:
              type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        finishedAt:
          type: [string, 'null']
          format: date-time
...
    ReportPreview:
      type: object
      required: [kind, generatedAt, notes, columns, items, page, size, total]
      properties:
        kind:
          $ref: '#/components/schemas/ReportKind'
        generatedAt:
          type: string
          format: date-time
        notes:
          type: array
          description: The same generation time, period rule, manager semantics and filter lines as in the file header
          items:
            type: string
        columns:
          type: array
          items:
            $ref: '#/components/schemas/ReportColumnView'
        items:
          type: array
          description: Row objects keyed by column id; dates are ISO 8601 strings, DEMAND measures are integers, missing values are null
          items:
            type: object
            additionalProperties:
              type: [string, integer, 'null']
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 200
        total:
          type: integer
          minimum: 0
...
    StatisticsRequest:
      type: object
      required: [kind, groupBy]
      properties:
        kind:
          $ref: '#/components/schemas/ReportKind'
        groupBy:
          $ref: '#/components/schemas/StatisticsGroupBy'
        from:
          type: [string, 'null']
          format: date
        to:
          type: [string, 'null']
          format: date
        periodBasis:
          oneOf:
            - $ref: '#/components/schemas/PeriodBasis'
            - type: 'null'
        filters:
          $ref: '#/components/schemas/ReportFilters'
...
    StatisticsResult:
      type: object
      required: [kind, groupBy, unit, timezone, generatedAt, filters, notes, total, unknownCount, items, maxChartBars]
      properties:
        kind:
          $ref: '#/components/schemas/ReportKind'
        groupBy:
          $ref: '#/components/schemas/StatisticsGroupBy'
        unit:
          type: string
        from:
          type: [string, 'null']
          format: date
        to:
          type: [string, 'null']
          format: date
        periodBasis:
          oneOf:
            - $ref: '#/components/schemas/PeriodBasis'
            - type: 'null'
        timezone:
          type: string
        generatedAt:
          type: string
          format: date-time
        filters:
          $ref: '#/components/schemas/ReportFilters'
        notes:
          type: array
          items:
            type: string
        total:
          type: integer
          minimum: 0
          description: >
            Number of report rows; for DEMAND the number of applications. PRODUCT counts may add up to more because one
            interaction can have several products
        unknownCount:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/StatisticsItem'
        maxChartBars:
          type: integer
          minimum: 1
          description: >
            Largest number of bars a PNG/PDF chart file can hold, counting the «Не указано» bar of DIRECTION, PROGRAM,
            PRODUCT and MANAGER; a larger chart job fails with REPORT_CHART_LIMIT
    ReportManagerOption:
      type: object
      required: [id, displayName, active]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        active:
          type: boolean
    SourceCode:
      type: string
      enum: [WEBSITE, MOODLE]
...
    SyncRunCreated:
      type: object
      required: [runId]
      properties:
        runId:
          type: string
          format: uuid
...
security:
  - cookieAuth: []